Separate oracle connectivity unit test to a separate spec file and disable for now; node:8-slim an use a pre-compiled oracledb add on, so go back to listing it directly as a dependency [skip ci]
