+++ package.json
@@ -52,6 +52,7 @@
     "express": "^4.16.3",
     "node-fetch": "^2.2.0",
     "node-ssh": "^5.1.2",
+    "oracledb": "^3.0.0",
     "shelljs": "^0.8.2",
     "ts-loader": "^5.2.1",
     "typescript": "^3.0.3",
